再次着手修复文件长度检查误报错问题。

+++ msxx_pak/reader.c
@@ -4,9 +4,9 @@
 bool PAKReader_file_check(PAKReader* restrict preader) // 检查文件
 {
 	const PAK_Item* lastItem = preader->itemTable[preader->itemCount - 1];
-	uint32_t fileLen = lastItem->relativeOffset + lastItem->fileLength, fileLen_real = get_file_len(preader->pak);
+	long long int fileLen = lastItem->relativeOffset + lastItem->fileLength, fileLen_real = get_file_len(preader->pak);
 
-	if (abs(fileLen - fileLen_real) < 4 * sizeof(uint32_t)) {
+	if (abs(fileLen - fileLen_real) < 4 * (long long int)sizeof(uint32_t)) {
 		return true;
 	} else {
 		return false;
